requirements.txt
asgiref==3.6.0
Django==4.1.5
django-taggit==3.0.0
django-widget-tweaks==1.4.12
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
